vertical_builder/requirements.txt
Flask>=3.0.0
gunicorn>=22.0.0
google-cloud-firestore>=2.20.0
google-cloud-storage>=2.16.0
